feat: add test for process payment activity
